+++ packages/psychart/src/psystate.ts
@@ -1,6 +1,6 @@
 import { SMath } from 'smath';
 import { PsyOptions, Datum, Layout, Point } from './types';
-const Psychrolib = require('./psychrolib');
+const Psychrolib = require('psychrolib');
 
 /**
  * Represents a single air condition using several states.
@@ -78,6 +78,12 @@ export class PsyState {
         PsyState.dbMin = config.dbMin;
         PsyState.dbMax = config.dbMax;
         PsyState.hrMax = Psychrolib.GetHumRatioFromTDewPoint(config.dpMax, PsyState.atm);
+    }
+    /**
+     * A static helper function to convert a humidity ratio into a dew point.
+     */
+    public static hr2dp(db: number, hr: number): number {
+        return Psychrolib.GetTDewPointFromHumRatio(db, hr, PsyState.atm);
     }
     /**
      * Initialize a new psychrometric state.