+++ packages/psychart/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+- Depend on npm package `psychrolib` instead of including its source code directly
+- Scale up humidity ratio by x1,000 and adjust units appropriately
+- Allow y-axis to be set to either dew point or humidity ratio (still allows flipping X and Y axes)
+- Update examples, show how to render a Mollier diagram with an example
+- Update base README file
+
 ## 0.1.1
 
 - Add defaults and incorporate into the example